+++ src/main/java/com/piveguyz/empickbackend/infra/s3/service/S3Service.java
@@ -1,5 +1,6 @@
-package com.piveguyz.empickbackend.infra.s3;
+package com.piveguyz.empickbackend.infra.s3.service;
 
+import com.piveguyz.empickbackend.infra.s3.dto.S3UploadResponseDTO;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.util.List;
@@ -8,7 +9,7 @@ public interface S3Service {
 
     byte[] download(String key);
 
-    String uploadFile(String prefix, String fileName, MultipartFile file);
+    S3UploadResponseDTO uploadFile(String prefix, String fileName, MultipartFile file);
 
     void delete(String key);
